requirements.txt
numpy>=2.2.3
pandas==2.2.3
scipy==1.15.2
connected-components-3d==3.23.0
alive-progress==3.2.0
matplotlib==3.10.1
opencv-python==4.11.0.86
torch==2.6.0
torchvision>=0.20.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
pillow>=9.4.0
iopath>=0.1.10
imageio>=2.9.0
opencv-python>=4.7.0
supervision==0.25.0
cupy-cuda12x==13.3.0
cucim-cu12==25.2.0